Add option for filtered list of countries
Add an optional parameter to specify a list of countries from the
location file (usually location.csv). The corresponding ipv4 and ipv6
subsets are written to additional files with "-interesting" suffix.

For example, to generate a subset of geoip sets only matching Spain,
France and Portugal addresses:

	$ ./nft_geoip.py -c es,fr,pt [...]
	[...]
	Found countries:  {'france': 'fr', 'portugal': 'pt', 'spain': 'es'}
	Writing interesting countries file...
	Done!

	$ ls -lah *-interesting.nft
	-rw-r--r-- ... geoip-ipv4-interesting.nft
	-rw-r--r-- ... geoip-ipv6-interesting.nft

The whole set geoip files are still generated.

In addition, add a parameter to print out the list of detected countries
from the location file:

	$ ./nft_geoip.py --file-location location.csv --show-countries

Before this commit, if any user wanted a subset of contries from
locations.csv it was needed to filter and trim the file before running
nft_geoip.py against the trimmed file. This is still possible, for
example:

	$ grep -e ES -e FR location.csv > locations_small.csv
	$ ./nft_geoip --file-location locations_small.csv [...]