Fall back to previous month before bailing out
It has been reported that DB-IP might take some days before releasing a
new dataset in the first days of a the month. It was proposed to add a
fallback to the previous month as to avoid the user tweaking the script
to fix this situation.

This commit adds said fallback logic, and only if previous month does
not download succesfully it will bail out.

Fixes #6